Relax requirements to match `ocean-dev:latest`
In examples (or repos that use `dwavesys/ocean-dev` container), it's
best to pin only what's actually necessary, and use looser specifiers
where possible, e.g. a lower bound only or a range.

This ensures older (but not actually required) versions are not
installed over the versions available in ocean-dev.

I.e. saving the planet one build at a time. :-)